Refactor database_helper.dart: Add isProductLikedByUser method
Refactored the database_helper.dart file to include a new method called isProductLikedByUser. This method checks if a product is liked by a user based on the product ID and user UID. It queries the 'likedProducts' table in the database and returns a boolean value indicating whether the product is liked or not.