Refactor product_card.dart: Update liked product creation
Refactor the _ProductCardState class in product_card.dart to update the creation of a liked product. Instead of manually passing individual properties, the LikedProduct.fromProduct method is now used to create a LikedProduct object from the widget's product. This improves code readability and maintainability.

Refactor database_helper.dart: Add nutriments field to likedProducts table schema

Add a new nutriments field to the likedProducts table schema in database_helper.dart. This field will store the nutritional information of the liked product. This enhancement allows for better tracking and analysis of user preferences.

Refactor liked_products.dart: Update LikedProduct class

Update the LikedProduct class in liked_products.dart to include the nutriments field. This field is now included in the toMap and fromMap methods, allowing for proper serialization and deserialization of the LikedProduct object. This change ensures consistency and compatibility with the updated likedProducts table schema.

Refactor qr_scanner_page.dart: Fetch product details before navigation

In qr_scanner_page.dart, update the navigation flow to fetch product details before navigating to the ProductDetailPage. This ensures that the product details are available when displaying the product information. The _getProduct method is introduced to fetch the product using the barcode, and the product is passed to the ProductDetailPage for display.

Refactor qr_scanner_page.dart: Handle error when fetching product

In qr_scanner_page.dart, handle errors that occur when fetching the product details. If an error occurs during the product fetch, the error is printed to the console. This provides better visibility and debugging information for any issues encountered during the product fetch process.